Renamed Navbar to BsNavbar
Bootstrap elements start with Bs prefix in order to easily differentiate them from our own components.
Plus relocated the USER ADD route to the top bar with a Sign Up button.
